+++ packages/_private/angular-demo/src/AllComponentParts/Input.ts
@@ -78,6 +78,17 @@ import { SynInputComponent } from '@synergy-design-system/angular/components/inp
       value="50"
     />
     <!-- /#838 -->
+
+    <!-- #872: Spin buttons out of sync with value -->
+    <syn-input
+      data-testid="input-872-spin-buttons"
+      label="Issue #872 - Spin Buttons out of sync with value"
+      [max]=100
+      [min]=0
+      type="number"
+      value="50"
+    />
+    <!-- /#872 -->
   `,
 })
 export class Input {}