+++ packages/components/src/components/input/input.component.ts
@@ -379,7 +379,10 @@ export default class SynInput extends SynergyElement implements SynergyFormContr
     }
 
     const min = typeof this.min === 'string' ? parseFloat(this.min) : this.min;
-    return this.valueAsNumber <= min;
+
+    // #872: Use value instead of valueAsNumber as valueAsNumber is drawn from the hidden input
+    // that is not yet updated in the current render cycle.
+    return parseFloat(this.value) <= min;
   }
 
   private isIncrementDisabled() {
@@ -392,7 +395,10 @@ export default class SynInput extends SynergyElement implements SynergyFormContr
     }
 
     const max = typeof this.max === 'string' ? parseFloat(this.max) : this.max;
-    return this.valueAsNumber >= max;
+
+    // #872: Use value instead of valueAsNumber as valueAsNumber is drawn from the hidden input
+    // that is not yet updated in the current render cycle.
+    return parseFloat(this.value) >= max;
   }